requirements.txt
Django==4.2.2
django_recaptcha==3.0.0
django_summernote==0.8.20.0
Markdown==3.4.3
